+++ crates/messaging/affinidi-messaging-mediator/affinidi-messaging-mediator-processors/src/lib/message_expiry_cleanup/database/scanning.rs
@@ -20,7 +20,7 @@ impl MessageExpiryCleanupProcessor {
 
         let mut conn = self.database.get_async_connection().await?;
 
-        deadpool_redis::redis::cmd("ZRANGE")
+        redis::cmd("ZRANGE")
             .arg("MSG_EXPIRY")
             .arg("-inf")
             .arg(now)
@@ -53,7 +53,7 @@ impl MessageExpiryCleanupProcessor {
         let mut total: u32 = 0;
 
         loop {
-            let msg_id: Option<String> = deadpool_redis::redis::cmd("SPOP")
+            let msg_id: Option<String> = redis::cmd("SPOP")
                 .arg(key)
                 .query_async(&mut conn)
                 .await
@@ -74,7 +74,7 @@ impl MessageExpiryCleanupProcessor {
                 }
                 total += 1;
             } else {
-                deadpool_redis::redis::cmd("ZREM")
+                redis::cmd("ZREM")
                     .arg("MSG_EXPIRY")
                     .arg(key)
                     .exec_async(&mut conn)