+++ crates/messaging/affinidi-messaging-mediator/src/database/acls.rs
@@ -25,7 +25,7 @@ impl Database {
 
             let mut con = self.get_connection().await?;
 
-            deadpool_redis::redis::Cmd::hset(
+            redis::Cmd::hset(
                 format!("DID:{did_hash}"),
                 "ACLS",
                 acls.to_hex_string(),
@@ -60,7 +60,7 @@ impl Database {
             let mut con = self.get_connection().await?;
 
             let acl: Option<String> =
-                deadpool_redis::redis::Cmd::hget(format!("DID:{did_hash}"), "ACLS")
+                redis::Cmd::hget(format!("DID:{did_hash}"), "ACLS")
                     .query_async(&mut con)
                     .await
                     .map_err(|err| {
@@ -154,7 +154,7 @@ impl Database {
         };
 
         if let Some(from_hash) = from_hash {
-            let (exists, acl): (bool, Option<String>) = match deadpool_redis::redis::pipe()
+            let (exists, acl): (bool, Option<String>) = match redis::pipe()
                 .atomic()
                 .cmd("SISMEMBER")
                 .arg(["ACCESS_LIST:", to_hash].concat())
@@ -233,7 +233,7 @@ impl Database {
             debug!("Requesting Access List");
 
             let mut con = self.get_connection().await?;
-            let (new_cursor, hashes): (u64, Vec<String>) = deadpool_redis::redis::cmd("SSCAN")
+            let (new_cursor, hashes): (u64, Vec<String>) = redis::cmd("SSCAN")
                 .arg(["ACCESS_LIST:", did_hash].concat())
                 .arg(cursor)
                 .arg("COUNT")
@@ -268,7 +268,7 @@ impl Database {
             debug!("Requesting Access List Count");
 
             let mut con = self.get_connection().await?;
-            deadpool_redis::redis::cmd("SCARD")
+            redis::cmd("SCARD")
                 .arg(["ACCESS_LIST:", did_hash].concat())
                 .query_async(&mut con)
                 .await
@@ -316,7 +316,7 @@ impl Database {
             };
 
             let mut con = self.get_connection().await?;
-            let mut query = deadpool_redis::redis::cmd("SADD");
+            let mut query = redis::cmd("SADD");
             let mut query = query.arg(["ACCESS_LIST:", did_hash].concat());
 
             for hash in hashes {
@@ -361,7 +361,7 @@ impl Database {
             debug!("Removing from Access List");
 
             let mut con = self.get_connection().await?;
-            let mut query = deadpool_redis::redis::cmd("SREM");
+            let mut query = redis::cmd("SREM");
             let mut query = query.arg(["ACCESS_LIST:", did_hash].concat());
 
             for hash in hashes {
@@ -391,7 +391,7 @@ impl Database {
             debug!("Clearing Access List");
 
             let mut con = self.get_connection().await?;
-            deadpool_redis::redis::cmd("DEL")
+            redis::cmd("DEL")
                 .arg(["ACCESS_LIST:", did_hash].concat())
                 .exec_async(&mut con)
                 .await
@@ -428,7 +428,7 @@ impl Database {
             debug!("Getting from Access List");
 
             let mut con = self.get_connection().await?;
-            let mut query = deadpool_redis::redis::cmd("SMISMEMBER");
+            let mut query = redis::cmd("SMISMEMBER");
             let mut query = query.arg(["ACCESS_LIST:", did_hash].concat());
 
             for hash in hashes {