+++ crates/messaging/affinidi-messaging-mediator/src/database/session.rs
@@ -159,7 +159,7 @@ impl Database {
 
         let sid = format!("SESSION:{}", session.session_id);
 
-        deadpool_redis::redis::pipe()
+        redis::pipe()
             .atomic()
             .cmd("HSET")
             .arg(&sid)
@@ -197,7 +197,7 @@ impl Database {
         let mut con = self.get_connection().await?;
 
         let (session_db, did_db): (HashMap<String, String>, Vec<Option<String>>) =
-            deadpool_redis::redis::pipe()
+            redis::pipe()
                 .atomic()
                 .cmd("HGETALL")
                 .arg(format!("SESSION:{session_id}"))
@@ -322,7 +322,7 @@ impl Database {
         let old_sid = format!("SESSION:{old_session_id}");
         let new_sid = format!("SESSION:{new_session_id}");
 
-        deadpool_redis::redis::pipe()
+        redis::pipe()
             .atomic()
             .cmd("RENAME")
             .arg(&old_sid)
@@ -365,7 +365,7 @@ impl Database {
 
         let sid = format!("SESSION:{session_id}");
 
-        deadpool_redis::redis::Cmd::hset(&sid, "refresh_token_hash", refresh_token_hash)
+        redis::Cmd::hset(&sid, "refresh_token_hash", refresh_token_hash)
             .exec_async(&mut con)
             .await
             .map_err(|err| {
@@ -388,7 +388,7 @@ impl Database {
 
         let sid = format!("SESSION:{session_id}");
 
-        let hash: Option<String> = deadpool_redis::redis::Cmd::hget(&sid, "refresh_token_hash")
+        let hash: Option<String> = redis::Cmd::hget(&sid, "refresh_token_hash")
             .query_async(&mut con)
             .await
             .map_err(|err| {